Consolidate backends with their invokers
