+++ src/backend_mmap.rs
@@ -5,8 +5,29 @@ use async_trait::async_trait;
 use bytes::{Buf, Bytes};
 use fmmap::tokio::{AsyncMmapFile, AsyncMmapFileExt as _, AsyncOptions};
 
-use crate::async_reader::AsyncBackend;
+use crate::async_reader::{AsyncBackend, AsyncPmTilesReader};
+use crate::cache::{DirectoryCache, NoCache};
 use crate::error::{PmtError, PmtResult};
+
+impl AsyncPmTilesReader<MmapBackend, NoCache> {
+    /// Creates a new `PMTiles` reader from a file path using the async mmap backend.
+    ///
+    /// Fails if [p] does not exist or is an invalid archive.
+    pub async fn new_with_path<P: AsRef<Path>>(path: P) -> PmtResult<Self> {
+        Self::new_with_cached_path(NoCache, path).await
+    }
+}
+
+impl<C: DirectoryCache + Sync + Send> AsyncPmTilesReader<MmapBackend, C> {
+    /// Creates a new cached `PMTiles` reader from a file path using the async mmap backend.
+    ///
+    /// Fails if [p] does not exist or is an invalid archive.
+    pub async fn new_with_cached_path<P: AsRef<Path>>(cache: C, path: P) -> PmtResult<Self> {
+        let backend = MmapBackend::try_from(path).await?;
+
+        Self::try_from_cached_source(backend, cache).await
+    }
+}
 
 pub struct MmapBackend {
     file: AsyncMmapFile,
